\name{chisqtest}
\alias{chisqtest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Calculate Chi-Square Tests
}
\description{
	Perform Pearson's Chi-Square Tests based on contingency table or perform batch 
	Pearson's Chi-Square Tests on a list of contingency tables.
}
\usage{
	ChisqTest(data);
	MultiChisqTest(data);
}
%- maybe also 'usage' for other objects documented here.
\details{
	ChisqTest(data);
	MultiChisqTest(data);
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
	Return P-value or a vector of P-values.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
	dengxiaodi
}
\note{
	More information about Chi-Square tests implementation. 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function () 
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
